Fix drop script
