+++ spring-integration-jdbc/src/main/resources/org/springframework/integration/jdbc/schema-drop-hsqldb.sql
@@ -3,4 +3,5 @@
 DROP TABLE  INT_MESSAGE IF EXISTS;
 DROP TABLE  INT_MESSAGE_GROUP IF EXISTS;
 DROP TABLE  INT_GROUP_TO_MESSAGE IF EXISTS;
+DROP TABLE  INT_LOCK IF EXISTS;
 DROP INDEX  INT_MESSAGE_IX1 IF EXISTS;
