+++ spring-integration-jdbc/src/main/resources/org/springframework/integration/jdbc/schema-drop-mysql.sql
@@ -3,4 +3,5 @@
 DROP TABLE IF EXISTS INT_MESSAGE ;
 DROP TABLE IF EXISTS INT_MESSAGE_GROUP ;
 DROP TABLE IF EXISTS INT_GROUP_TO_MESSAGE ;
+DROP TABLE IF EXISTS INT_LOCK ;
 DROP INDEX IF EXISTS INT_MESSAGE_IX1 ;
