+++ spring-integration-jdbc/src/main/resources/org/springframework/integration/jdbc/schema-drop-postgresql.sql
@@ -3,4 +3,5 @@
 DROP TABLE  INT_MESSAGE ;
 DROP TABLE  INT_MESSAGE_GROUP ;
 DROP TABLE  INT_GROUP_TO_MESSAGE ;
+DROP TABLE  INT_LOCK ;
 DROP INDEX  INT_MESSAGE_IX1 ;
