+++ spring-integration-jdbc/src/main/sql/destroy.sql.vpp
@@ -1,4 +1,5 @@
 DROP TABLE $!{IFEXISTSBEFORE} INT_MESSAGE $!{IFEXISTS};
 DROP TABLE $!{IFEXISTSBEFORE} INT_MESSAGE_GROUP $!{IFEXISTS};
 DROP TABLE $!{IFEXISTSBEFORE} INT_GROUP_TO_MESSAGE $!{IFEXISTS};
+DROP TABLE $!{IFEXISTSBEFORE} INT_LOCK $!{IFEXISTS};
 DROP INDEX $!{IFEXISTSBEFORE} INT_MESSAGE_IX1 $!{IFEXISTS};
